New feature - 'early night'
If the alarm is set by a scheduled task, the code first checks if the
alarm is already set.
If the alarm isn't already set, an email is sent.
If the alarm is already set, the event is logged, but no email is sent.
This means setting the alarm manually will prevent emails being sent by
a scheduled task.